Create code
This function generates a random hex color code, which can be used to set the background or text color of an HTML element. The function uses a loop to generate a string of 6 random characters, each of which corresponds to a value between 0 and 15 in the hexadecimal number system. The resulting string is prepended with a '#' symbol to create a valid color code.